Compatibility with python3.
Changed long to int and removed the auto field for the fit model since sqlite3
was complaining when leaving NULL for the pk. Django auto-creates this field for
us.

+++ ui/views/index.py
@@ -78,7 +78,7 @@ def __getExpandedGroups(request):
   expandedGroups = []
   if "expandedMarketGroups" in request.COOKIES:
     cookie = json.loads(request.COOKIES["expandedMarketGroups"])
-    expandedGroups = [long(x) for x in cookie]
+    expandedGroups = [int(x) for x in cookie]
 
   return expandedGroups
